pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Denise Grassi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.4.1"
django = "^5.2.3"
djangorestframework = "^3.16.0"
python-dotenv = "^1.0.1"
pytest-django = "^4.11.1"
faker = "^37.4.0"
factory-boy = "^3.3.3"
django-extensions = "^4.1"
django-debug-toolbar = "^5.2.0"
psycopg2-binary = "^2.9.10"
flake8 = "^7.3.0"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
